adds further documentation to master data

+++ src/main/java/api/masterData/loadingUnit/Container.java
@@ -1,7 +1,21 @@
 package api.masterData.loadingUnit;
 
 /**
- * Container for the transport of goods for the change of another mean of transport.
+ * Case for the transport of goods for the change of another mean of transport.
+ *
+ * <p>DIGIT_name: Container</p>
+ *
+ * <p>DIGIT_english: Container</p>
+ *
+ * <p>DIGIT_definition: Behälter für den Gütertransport für den Wechsel von einer zur anderen Transportform.</p>
+ *
+ * <p>DIGIT_definition_english: Case for the transport of goods for the change to another means of transport.</p>
+ *
+ * <p>DIGIT_annotation: Containertypen können Flat-, Open-Top-, High-Cube-, Bulk-, ISO-, Reefer- oder Tank-Container
+ * sein. (@see DIN EN ISO 17261:2012-12, 3.14 — modifiziert, Änderung der Anmerkung)</p>
+ *
+ * <p>DIGIT_annotation_english: Container types can be flat, open top, high cube, bulk, ISO, reefer- or tank
+ * containers. (@see DIN EN ISO 17261:2012-12, 3.14 — modifiziert, Änderung der Anmerkung)</p>
  *
  * @author  Isabell Dürlich - duerlich@synyx.de
  */
@@ -12,11 +26,13 @@ public class Container extends LoadingUnit {
      */
     private String sizeType;
 
-    // TODO - textuelle beschreibung zu size type?
+    /**
+     * Container types can be flat, open top, high cube, bulk, ISO, reefer- or tank containers.
+     */
     private String type;
 
     /**
      * in foot.
      */
-    private double size;
+    private int size;
 }
